ScoreLens_FastApi/requirements.txt
fastapi
uvicorn[standard]
python-multipart
confluent-kafka
kafka-python
python-dotenv

#asw s3
boto3

#postgres
sqlalchemy  #ORM giống hybernate
psycopg2-binary #postgre driver
alembic #quản lý migrate DB schema


#alembic init alembic

#khi có model mới hoặc sửa model
#set PYTHONPATH=.
#alembic revision --autogenerate -m "create balls table"

#tạo db
#alembic upgrade head

